feat: implement role-based access control; add authentication middleware and user role checks; enhance event creation logic

+++ src/services/user.service.ts
@@ -1,6 +1,7 @@
+import bcrypt from "bcrypt";
+import { isValidObjectId } from "mongoose";
 import { UserInterface } from "../interfaces";
 import { UserModel } from "../models/user.model";
-import bcrypt from "bcrypt";
 
 export class UserService {
   async createUser(name: string, email: string, password: string) {
@@ -25,14 +26,50 @@ export class UserService {
     return await UserModel.findById(id);
   }
 
-  async updateUser(id: string, userData: UserInterface) {
-    if(userData.password){
+  async updateUserPartial(id: string, updateData: UserInterface) {
+    return await UserModel.findByIdAndUpdate(id, updateData, { new: true });
+  }
+
+  async updateUserFull(id: string, userData: UserInterface) {
+    const existingUser = await UserModel.findById(id);
+    if (!existingUser) {
+      throw new Error("Utilisateur introuvable ❌");
+    }
+
+    if (userData.password) {
       userData.password = await bcrypt.hash(userData.password, 10);
     }
-    return await UserModel.findByIdAndUpdate(id, userData, { new: true });
+
+    return await UserModel.findByIdAndUpdate(id, userData, {
+      new: true,
+      overwrite: true,
+    });
   }
 
   async deleteUser(id: string) {
-    return await UserModel.findByIdAndDelete(id);
+    if (!isValidObjectId(id)) {
+      throw new Error("ID utilisateur invalide");
+    }
+
+    const deletedUser = await UserModel.findByIdAndDelete(id);
+    if (!deletedUser) {
+      throw new Error("Utilisateur non trouvé");
+    }
+  }
+
+  async createSuperAdminIfNotExists() {
+    const existingSuperAdmin = await UserModel.findOne({ role: "superadmin" });
+
+    if (!existingSuperAdmin) {
+      const hashedPassword = await bcrypt.hash("didipurple", 10); // 🛑 Change ce mot de passe après installation
+      await UserModel.create({
+        name: "dylane",
+        email: "dylane@didicode.com",
+        password: hashedPassword,
+        role: "superadmin",
+      });
+
+      console.log("✅ Super Admin créé avec succès !");
+    }
   }
 }
